Select Diverging Simulations from an sim_rw Object

The selection will contain the most extreme simulations base on either the minimum effect or the maximum effect within the simulation.

Usage

select_divergence(x, n = 10)

Arguments

x

An sim_rw object. Which is the output of \link{simulate_rw}

n

The number of simulations to plot when only a subset is shown.

See also